Complete working implementation of async FlowIR
Finally able to get wavenumbers from spectra! Yay!
It was a property of the opc standard datatype YArrayItem,
and the access via asyncua was not the most intuitive...

A new class Spectrum has been created (likely to become
abstract base class for all analytic stuff) from which
SpectrumIR inherits.

TODO: integrals in spectrum base class and the script
to follow reaction yield over time with a realtime
plot (WIP stub on MFC PC not yet committed).