Further improvement to peak fitting.
Now, you might think IR bands are Gaussian. Or Lorentzian.
And sometime you'd be right. But also oh, so wrong![1]
Reality is that life's complicated, and they can show both Gaussian
and Lorentzian character at the same time.
So let's fit with a PseudoVoigtModel, preliminary test shows it is a lot better!
I believe we are at a point where we can cross-validate a couple of IR-yields w/ NMR and that's it for this reaction analysis!

[1] As a primer see. You'll realize that the Gaussian-Lorentzia is actually one of the simplest fitting approach :D https://assets.thermofisher.com/TFS-Assets/CAD/Application-Notes/AN50733_E.pdf