Base on the historical changes and the commits:
17a83a868b836ab44aa8e99bb66d207eb4710a47

934be6bdb4e2cba5745478f791de602f31db2375

ed9444e005f8737e7134b0dbff65f00c1a06aeed

3a3b9393d994c38f288256bffc9aa331b538d010

The valve test failed repeatedly in two methods.

This failure only occurred when using the get_all_flowchem_devices() approach. The method works on the server (interacting with the device through the browser).

The root cause is that the argument encoding the list in a way that FastAPI isn't handling by default. The most appropriate solution is to work with strings.