Solution of the issue 141
Enforce overwriting of parent methods, dynamic creation of API nodes.
This work perfectly. I just want to schedule a meting with the team to see the viability of merge it!
If you see advantage in this new feature, it will be necessary to add one line of code in every component in flowchem!
The discussion about the difference between GET and PUT remains. I mean, if it going to be implement, the name of some methods in the API server will be the same of the class.