Creating a new device to monitoring and recording the datas - chronologically
The best way to start to build a supervisor system and integrate it in the flowchem is creating a new device (virtual device). This device will access all attributes of the Flowchem package to create a file of information about the experimental. The user will access it directly in the API server in order to insert the path and the experimental identification.
It was the best way I find to build the monitoring and watchdog.